move everything back to root
