+++ .storybook/components/SpacedGroup.d.ts
@@ -1,4 +1,4 @@
-import { FlameFlexProps } from '../../src/Core';
+import { FlameFlexProps } from '../../packages/flame/src/Core';
 
 export interface SpacedGroupProps extends FlameFlexProps {}
 export const SpacedGroup: React.FC<SpacedGroupProps>;
