+++ .storybook/preview-head.html
@@ -8,6 +8,7 @@
       display: none;
     }
   }
+  html,
   body {
     font-family: Lato, Helvetica Neue, Helvetica, Arial, sans-serif;
   }
