+++ .storybook/stories.scss
@@ -1,4 +1,5 @@
-@import '../../packages/flame-tokens/partials/index.scss';
+@import '../packages/flame-tokens/partials/index.scss';
+@import '../packages/flame-css/src/index.scss';
 
 // Required imports for state placeholders use
