+++ package.json
@@ -19,7 +19,7 @@
     "test:ci": "yarn test --coverage --runInBand && codecov",
     "typecheck": "tsc",
     "precommit": "lint-staged",
-    "dev": "yarn && concurrently --kill-others 'yarn dev:tokens' 'yarn dev:themes' 'lerna exec --scope \"@lightspeed/flame\" yarn storybook'",
+    "dev": "yarn && concurrently --kill-others 'yarn dev:tokens' 'yarn dev:themes' 'yarn start-storybook -p 6006'",
     "dev:themes": "chokidar './packages/flame/themes/**/*.ts' -c 'lerna run --scope @lightspeed/flame build:themes'",
     "dev:tokens": "chokidar './packages/flame-tokens/src/*.ts' -c 'lerna run --scope @lightspeed/flame-tokens prepublish'",
     "build:storybook": "yarn bootstrap && lerna exec --scope \"@lightspeed/flame\" yarn build-storybook",
@@ -31,7 +31,7 @@
     "release-and-publish": "yarn release && yarn lerna:publish && lerna run --scope @lightspeed/flame release",
     "generate": "plop",
     "docgen": "node ./scripts/docgen",
-    "percy": "lerna exec --scope \"@lightspeed/flame\" yarn percy"
+    "percy": "build-storybook -c .storybook && percy-storybook --widths=1280"
   },
   "dependencies": {
     "@emotion/core": "^10.0.9",
@@ -57,6 +57,11 @@
     "@lightspeed/config-prettier": "^0.1.2",
     "@lightspeed/config-typescript": "^0.2.0",
     "@lightspeed/eslint-config": "^1.1.1",
+    "@percy/storybook": "^3.2.0",
+    "@storybook/addon-actions": "^5.3.19",
+    "@storybook/addon-docs": "^5.3.19",
+    "@storybook/addons": "^5.3.19",
+    "@storybook/react": "^5.3.19",
     "@testing-library/react": "^10.0.5",
     "@testing-library/jest-dom": "^5.9.0",
     "@types/classnames": "^2.2.7",
@@ -93,6 +98,10 @@
     "postcss-loader": "^2.1.5",
     "prettier": "^1.16.1",
     "react-test-renderer": "^16.8.6",
+    "sass": "^1.14.3",
+    "sass-loader": "^7.1.0",
+    "style-loader": "^0.21.0",
+    "storybook-readme": "^5.0.8",
     "typescript": "^3.5.3",
     "webpack": "^4.41.2",
     "webpack-cli": "^2.1.3",
