+++ packages/flame-css/package.json
@@ -11,21 +11,10 @@
   },
   "scripts": {
     "build": "node ./scripts/src-to-css.js",
-    "dev": "yarn storybook",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook",
     "prepublish": "yarn build"
   },
   "devDependencies": {
-    "@babel/core": "7.10.3",
-    "@storybook/addon-docs": "^5.3.19",
-    "@storybook/react": "^5.3.19",
-    "autoprefixer": "^9.8.2",
-    "babel-loader": "8.1.0",
     "fs-extra": "9.0.1",
-    "postcss-import": "^12.0.1",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
     "sass": "^1.26.9"
   }
 }
