+++ packages/flame-css/stories/components.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Preview } from '@storybook/addon-docs/blocks';
 
-<Meta title="Components" />
+<Meta title="Flame-CSS|Components" />
 
 # Components
