--- packages/flame/.storybook/manager-head.html
@@ -1,1 +1,0 @@
-<meta name="viewport" content="width=device-width, initial-scale=1">
