--- packages/flame/.storybook/preview-head.html
@@ -1,11 +1,0 @@
-<link
-  href="https://fonts.googleapis.com/css?family=Lato:400,700&subset=latin-ext"
-  rel="stylesheet"
-/>
-<style>
-  @media only percy {
-    .hide-in-percy {
-      display: none;
-    }
-  }
-</style>
