+++ packages/flame/package.json
@@ -11,7 +11,6 @@
     "url": "https://github.com/lightspeed/flame/issues"
   },
   "scripts": {
-    "storybook": "start-storybook -p 6006",
     "bootstrap": "yarn build:icons && yarn build:flags && yarn build:themes",
     "prebuild": "rimraf ./dist",
     "run-svgo": "node ./scripts/run-svgo.js",
@@ -54,45 +53,18 @@
     "type-fest": "^0.3.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.1.6",
-    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.1.0",
     "@lightspeed/flame-tokens": "^1.0.0",
-    "@percy/storybook": "^3.2.0",
-    "@storybook/addon-actions": "^5.3.19",
-    "@storybook/addons": "^5.3.19",
-    "@storybook/react": "^5.3.19",
     "@types/lodash": "^4.14.123",
-    "@types/react": "^16.8.23",
     "@types/react-modal": "^3.8.1",
     "@types/react-select": "^2.0.15",
-    "@types/storybook-readme": "^5.0.0",
     "@types/styled-system__css": "^5.0.11",
     "@types/styled-system__theme-get": "^5.0.0",
-    "autoprefixer": "^8.0.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.5.0",
-    "babel-loader": "^8.0.4",
-    "babel-plugin-emotion": "^10.0.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.13",
-    "concurrently": "^3.5.1",
-    "css-loader": "^0.28.11",
-    "file-loader": "^1.1.11",
     "fs-extra": "7.0.1",
     "glob": "7.1.3",
     "htmltojsx": "^0.3.0",
     "lodash": "^4.17.11",
-    "prettier": "^1.16.1",
-    "raw-loader": "^0.5.1",
     "react-docgen-typescript": "^1.15.1",
     "rimraf": "2.6.2",
-    "sass": "^1.14.3",
-    "sass-loader": "^7.1.0",
-    "storybook-readme": "^5.0.8",
-    "style-loader": "^0.21.0",
     "ts-node": "8.3.0"
   },
   "publishConfig": {
