+++ packages/flame/src/Tooltip/story.tsx
@@ -4,7 +4,7 @@ import { withReadme } from 'storybook-readme';
 
 import { Tooltip, TooltipPlacement } from './Tooltip';
 import { Input } from '../Input';
-import { ExampleBox } from '../../.storybook/components/ExampleBox';
+import { ExampleBox } from '../../../../.storybook/components/ExampleBox';
 import Readme from './README.md';
 
 const stories = storiesOf('Components|Tooltip', module).addDecorator(withReadme(Readme));
